Source code release for version 9.1.0.0
on-behalf-of: @alternetsoft <yevgeni.zolotko@alternetsoft.com>

+++ Resources/Scripter.IronPython/CallMethod.py
@@ -9,14 +9,14 @@ def OnPaint(g, bounds, currentAngle):
   maxSide = Math.Max(bounds.Width, bounds.Height)
   radius = (maxSide - (maxSide / 3)) / 2
   dotRadius = maxSide / 20
-  center = Point(bounds.Left + bounds.Width / 2, bounds.Top + bounds.Height / 2)
+  center = Point(int(bounds.Left + bounds.Width / 2), int(bounds.Top + bounds.Height / 2))
   g.SmoothingMode = System.Drawing.Drawing2D.SmoothingMode.HighQuality
   g.Clear(SystemColors.Control)
   arenaBounds = bounds
   arenaBounds.Inflate(-2, -2)
   g.FillEllipse(arenaBackgroundBrush, arenaBounds)
   radians = DegreesToRadians(currentAngle)
-  dotCenter = Point(center.X + (Math.Cos(radians) * radius), center.Y + (Math.Sin(radians) * radius))
+  dotCenter = Point(int(center.X + (Math.Cos(radians) * radius)), int(center.Y + (Math.Sin(radians) * radius)))
   g.FillEllipse(dotBrush, dotCenter.X - dotRadius, dotCenter.Y - dotRadius, dotRadius * 2, dotRadius * 2)
 
 def ConstrainAngle(x):
