+++ Resources/Scripter.IronPython/ThreadScriptWPF.py
@@ -87,7 +87,7 @@ def DoQuickSort(iLo, iHi, ct):
   YPos2 = 0
   len1 = 0
   len2 = 0
-  mid = quickArray[((lo + hi) / 2)]
+  mid = quickArray[int(((lo + hi) / 2))]
   while lo <= hi:
     if ct.IsCancellationRequested:
         ct.ThrowIfCancellationRequested()
